+++ tests/test_codestyle.py
@@ -1,19 +1,8 @@
+#
 # Copyright (c) nexB Inc. and others.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Visit https://aboutcode.org and https://github.com/nexB/ for support and download.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 
 import subprocess
 import unittest
@@ -21,8 +10,15 @@ import unittest
 
 class BaseTests(unittest.TestCase):
     def test_codestyle(self):
-        args = "python -m black --check -l 100 ."
+        args = "venv/bin/black --check -l 100 setup.py etc src tests docs"
         try:
             subprocess.check_output(args.split())
-        except Exception as e:
-            raise Exception(f"Black style check failed, please format the code using black") from e
+        except subprocess.CalledProcessError as e:
+            print("===========================================================")
+            print(e.output)
+            print("===========================================================")
+            raise Exception(
+                "Black style check failed; please format the code using:\n"
+                "  python -m black -l 100 setup.py etc src tests docs",
+                e.output,
+            ) from e