+++ tests/test_gentoo.py
@@ -1,6 +1,8 @@
 # test_vercmp.py -- Portage Unit Testing Functionality
 # Copyright 2006 Gentoo Foundation
-# Distributed under the terms of the GNU General Public License v2
+# SPDX-License-Identifier: GPL-2.0-only
+# this has been significantly modified from the original
+
 
 from unittest import TestCase
 from univers.gentoo import vercmp